Move get_starcheck_obs_kwargs to tests.test_common
